Initial data combining and start of analysis
This script summarizes the raw csv HIMS files and outputs an inflow-outflow summary (annual_summary.csv) as well as a summary of program volume and exit destinations (program_summary.csv).  Note that the 'Initial_Count' field in the inflow-outflow table is calculated incorrectly because its using enrollments rather than unique client ids.  There are also some issues that need to be clarified with a subject matter expert. See the python notebook for details on how measures were derived.